requirements.txt
streamlit==1.41.1
pandas==2.2.3
requests==2.32.3
pypdf==4.3.1
Pillow==10.4.0
pytesseract==0.3.13
python-docx==1.1.2
